+++ js/src/components/ImageWithCaptionThumbnail.js
@@ -1,13 +1,27 @@
 import React from 'react';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
-import { styled } from '@material-ui/core';
+import { styled, withStyles } from '@material-ui/core';
 import LessonGrid from '@anu/components/LessonGrid';
 
 const StyledImg = styled('img')({
   width: '100%',
   display: 'block',
 });
+
+const StyledTypography = withStyles((theme) => ({
+  root: {
+    '& > p': {
+      marginBottom: theme.spacing(4),
+    },
+    '& > p:first-child': {
+      marginTop: 0,
+    },
+    '& > p:last-child': {
+      marginBottom: 0,
+    },
+  },
+}))(Typography);
 
 const ImageWithCaptionThumbnail = ({ image, caption }) => (
   <LessonGrid>
@@ -17,7 +31,7 @@ const ImageWithCaptionThumbnail = ({ image, caption }) => (
       </Grid>
 
       <Grid item md={8} xs={12}>
-        <Typography>{caption}</Typography>
+        <StyledTypography dangerouslySetInnerHTML={{ __html: caption }} />
       </Grid>
     </Grid>
   </LessonGrid>