+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Adds text formatting ability for paragraph types: "Image with caption (thumbnail)" and "List"
+
 ## [2.3.0]
 
  - Adds offline mode support for using with Progressive Web App module